FlaskPortfolioPort/requirements.txt
blinker==1.4
Flask==0.10.1
Flask-Admin==1.4.0
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Security==1.7.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.12
itsdangerous==0.24
Jinja2==2.8
lxml==3.6.0
MarkupSafe==0.23
ordereddict==1.1
passlib==1.6.5
Pillow==3.1.1
requests==2.10.0
requests-cache==0.4.12
SQLAlchemy==1.0.12
Werkzeug==0.11.5
WTForms==2.1
